unit tests for error u8 and u16 values
