+++ Cargo.toml
@@ -34,6 +34,7 @@ base58 = {version = "0.2.0", default-features = false}
 bitflags = {version = "1.3.2", default-features = false}
 cfg-if = "1.0"
 clap = "4.1.6"
+enum-iterator = "2.0.0"
 env_logger = "0.10.0"
 fluent-uri = { version = "0.1.4", default-features = false }
 futures = {version = "0.3.21", default-features = false}
