+++ crates/kilt-dip-primitives/src/merkle/v0/mod.rs
@@ -17,6 +17,7 @@
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
 //! Module to deal with cross-chain Merkle proof as generated by the KILT chain.
+
 use did::{
 	did_details::{DidPublicKey, DidPublicKeyDetails},
 	DidSignature, DidVerificationKeyRelationship,
@@ -35,13 +36,16 @@ use sp_std::{fmt::Debug, vec::Vec};
 use sp_trie::{verify_trie_proof, LayoutV1};
 
 use crate::{
-	state_proofs::{verify_storage_value_proof, verify_storage_value_proof_with_decoder, MerkleProofError},
+	state_proofs::{verify_storage_value_proof, verify_storage_value_proof_with_decoder},
 	traits::{BenchmarkDefault, GetWithArg},
 	utils::{
 		calculate_dip_identity_commitment_storage_key_for_runtime, calculate_parachain_head_storage_key,
 		BoundedBlindedValue, OutputOf,
 	},
 };
+
+mod error;
+pub use error::*;
 
 /// The state proof for a parachain head.
 ///
@@ -184,49 +188,6 @@ where
 		Self {
 			signature: DidSignature::worst_case(context),
 			valid_until: BlockNumber::default(),
-		}
-	}
-}
-
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, TypeInfo)]
-pub enum Error {
-	InvalidRelayHeader,
-	RelayBlockNotFound,
-	RelayStateRootNotFound,
-	InvalidDidMerkleProof,
-	TooManyLeavesRevealed,
-	InvalidSignatureTime,
-	InvalidDidKeyRevealed,
-	ParaHeadMerkleProof(MerkleProofError),
-	DipCommitmentMerkleProof(MerkleProofError),
-	Internal,
-}
-
-impl From<Error> for u8 {
-	fn from(value: Error) -> Self {
-		match value {
-			// DO NOT USE 0
-			// Errors of different sub-parts are separated by a `u8::MAX`.
-			// A value of 0 would make it confusing whether it's the previous sub-part error (u8::MAX)
-			// or the new sub-part error (u8::MAX + 0).
-			Error::InvalidRelayHeader => 1,
-			Error::RelayBlockNotFound => 2,
-			Error::RelayStateRootNotFound => 3,
-			Error::InvalidDidMerkleProof => 4,
-			Error::TooManyLeavesRevealed => 5,
-			Error::InvalidSignatureTime => 6,
-			Error::InvalidDidKeyRevealed => 7,
-			Error::ParaHeadMerkleProof(error) => match error {
-				MerkleProofError::InvalidProof => 11,
-				MerkleProofError::RequiredLeafNotRevealed => 12,
-				MerkleProofError::ResultDecoding => 13,
-			},
-			Error::DipCommitmentMerkleProof(error) => match error {
-				MerkleProofError::InvalidProof => 21,
-				MerkleProofError::RequiredLeafNotRevealed => 22,
-				MerkleProofError::ResultDecoding => 23,
-			},
-			Error::Internal => u8::MAX,
 		}
 	}
 }
